Skipping the test on Ubuntu for now
It looks like the parallel execution of the tests influences the server
and therefore makes the test unreliable. For example, the test on Linux
may create a new data entry and parallel running test on Windows
receives the new data comparing it with the expected data but now we
have one entry to much and the test fails.

Long term solution: Include Linux in the matrix build, we only need to
figure out how to make the build comand different for Linux and Win /
Mac